logout

+++ client/js/components/trip_list.js
@@ -1,28 +1,37 @@
-function renderTripList () {
+function renderTripList() {
   document.querySelector('#page').innerHTML = `
     <section class="trip-list">
-    ${renderTrips()}
+      ${renderTrips()}
     </section>
   `
 }
 
 function renderTrips() {
-    return state.trips.map(trip =>`
-  <section class='trip' data-id='${trip.id}'>
-    <header>
-      <h2${trip.name}</h2>
-      <span onClick="deleteTrip(event)">delete</span>
-      <span onClick="renderEditTrip()">edit</span>
-    </header>
-    <p>${trip.start_date}</p>
-    <p>${trip.end_date}</p>
-  </section>
-  `).join('')
+  if (state.loggedInUserName) {
+    console.log('logged in')
+    return state.trips.map(trip => `
+    <section class='trip' data-id='${trip.id}'>
+      <header>
+        <h2>${trip.name}</h2>
+        <span onClick="deleteTrip(event)">delete</span>
+        <span onClick="renderEditTrip()">edit</span>
+      </header>
+      <p>${trip.start_date}</p>
+      <p>${trip.end_date}</p>
+    </section>
+    `).join('')
+  } else {
+    console.log('not logged in')
+    console.log(state.trips)
+    return state.trips.map(trip => `
+    <section class='trip' data-id='${trip.id}'></section>
+    `).join('')
+  }
 }
 
 function renderEditTrip() {
   return state.trips.map(trip =>
-  document.querySelector('#page').innerHTML =`
+    document.querySelector('#page').innerHTML = `
   <section class='edit-trip' data-id='${trip.id}'>
     <form onSubmit="editTrip(event)">
       <h2>Edit Trip</h2>
@@ -47,19 +56,19 @@ function renderEditTrip() {
 }
 
 function editTrip(event) {
-  event.preventDefault() 
+  event.preventDefault()
   const form = event.target
   const data = Object.fromEntries(new FormData(form))
   fetch('/api/trips', {
     method: 'PUT',
-    headers: { "Content-Type": "application/json"}, 
+    headers: { "Content-Type": "application/json" },
     body: JSON.stringify(data)
   })
-  .then(res => res.json())
-  .then(trip => {
-    state.trips.push(trip)
-    renderTripList()
-  })
+    .then(res => res.json())
+    .then(trip => {
+      state.trips.push(trip)
+      renderTripList()
+    })
 }
 
 function deleteTrip(event) {
